+++ R/durations.r
@@ -215,6 +215,9 @@ duration <- function(num = NULL, units = "seconds", ...){
 }
 
 .duration_from_num <- function(num, units){
+  if(isS4(num) || !is.numeric(num)){
+    stop(sprintf("First argument to `duration` constructor must be character or numeric. Supplied object of class '%s'", class(num)))
+  }
   unit <- standardise_date_names(units)
   mult <- c(second = 1, minute = 60, hour = 3600, mday = 86400,
     wday = 86400, yday =86400, day = 86400, week = 604800,