+++ NEWS.md
@@ -30,6 +30,7 @@ Version 1.5.6.9000 (development)
 * [#456](https://github.com/hadley/lubridate/issues/456) Evaluation output in documentation examples was removed.
 
 ### BUG FIXES
+* [#462](https://github.com/hadley/lubridate/issues/462) Non-numeric or non-character arguments are disallowed as arguments to `period` and `duration` constructors.
 * [#458](https://github.com/hadley/lubridate/issues/458) When year is missing in parsing, return consistently year 0.
 * [#448](https://github.com/hadley/lubridate/issues/448) Correctly handle missing months and days in C parser.
 * [#450](https://github.com/hadley/lubridate/issues/450) Fix incorrect handling of DST gaps in `date_decimal` and `decimal_date`.